+++ timeutil/CHANGELOG.md
@@ -8,7 +8,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 ### Added
 
-* Add backoff and retry packages.
+* Add backoff, clock, and retry packages.
 
 ### Changed